import pooch_cache from conftest
