+++ src/openfe/tests/protocols/restraints/test_geometry_utils.py
@@ -3,12 +3,10 @@
 
 import itertools
 import os
-import pathlib
 from importlib import resources
 
 import MDAnalysis as mda
 import numpy as np
-import pooch
 import pytest
 from openff.units import unit
 from rdkit import Chem
@@ -32,7 +30,7 @@ from openfe.protocols.restraint_utils.geometry.utils import (
     stable_secondary_structure_selection,
 )
 
-from ...conftest import HAS_INTERNET
+from ...conftest import HAS_INTERNET, POOCH_CACHE
 
 
 @pytest.fixture(scope="module")
@@ -47,29 +45,6 @@ def eg5_protein_ligand_universe(eg5_protein_pdb, eg5_ligands):
     # add the residue name of the ligand
     lig.add_TopologyAttr("resname", ["LIG"])
     return mda.Merge(protein.atoms, lig.atoms)
-
-
-POOCH_CACHE = pooch.os_cache("openfe")
-zenodo_restraint_data = pooch.create(
-    path=POOCH_CACHE,
-    base_url="doi:10.5281/zenodo.15212342",
-    registry={
-        "t4_lysozyme_trajectory.zip": "sha256:e985d055db25b5468491e169948f641833a5fbb67a23dbb0a00b57fb7c0e59c8"
-    },
-)
-
-
-@pytest.fixture
-def t4_lysozyme_trajectory_universe():
-    zenodo_restraint_data.fetch("t4_lysozyme_trajectory.zip", processor=pooch.Unzip())
-    cache_dir = pathlib.Path(
-        pooch.os_cache("openfe") / "t4_lysozyme_trajectory.zip.unzip/t4_lysozyme_trajectory"
-    )
-    universe = mda.Universe(
-        str(cache_dir / "t4_toluene_complex.pdb"),
-        str(cache_dir / "t4_toluene_complex.xtc"),
-    )
-    return universe
 
 
 @pytest.fixture
